Seperate the duration changer view into functions
The entire view was a single 100+ line function, so I ended up seperating
it into 3 functions: a click event handler, a input element, and a public
one that joins the two together.

Also, why does any variable only used inside a macro show up as "unused"
by rust analyzer? This is annoying, as I need to prefix a lot of variables
with an underscore so that I don't get a ton of warnings.